flask-microservice/requirements.txt
flask==3.0.3
joblib==1.4.2
scikit-learn==1.7.1
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1